Add function to clean data
