+++ src/utils.js
@@ -1,0 +1,19 @@
+const cleanData = data => {
+    const cleanedData = data.map(movie => {
+      const splitDate = movie.release_date.split('-')
+      let [year, month, day] = splitDate
+      const releaseDateFormat = [month, day, year].join('-')
+
+      return {
+        id: movie.id,
+        poster_path: movie.poster_path,
+        backdrop_path: movie.backdrop_path,
+        title: movie.title,
+        average_rating: movie.average_rating.toFixed(2),
+        release_date: releaseDateFormat
+      }
+  })
+  return cleanedData;
+}
+
+export default cleanData
